Change the dialog files
